+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0] - 08-10-2025
+
+### Added
+- Example result initialized on db creation. This example can be directly retrieved from the FE.
+
+### Changed
+- JS scripts moved to individual /static/scripts
+
 ## [2.3.0] - 06-10-2025
 
 ### Changed